requirements.txt
asgiref==3.5.2
aspose-cells-python==22.8.0
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
dj-email-url==1.0.5
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-cache-url==3.4.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-spectacular==0.23.1
environs==9.5.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
jsonschema==4.13.0
marshmallow==3.17.0
oauthlib==3.2.0
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
